requirements.txt
requests
beautifulsoup4
psycopg2-binary
selenium
python-dotenv
pandas
sqlalchemy
alembic
requests-html
gunicorn
apache-airflow
